feat(tests): Fastapp auth unit tests moved to a common auth unit test including launch and run code
- Implement test_common_unit_auth method in run.py to execute unit tests for authentication
- Create test_auth.py in common auth unit that inherits from fastapp auth unit and uses the same set up code
