+++ .vscode/launch.json
@@ -568,6 +568,17 @@
             "justMyCode": true
         },
         {
+            "name": "test.common.unit.auth",
+            "type": "debugpy",
+            "request": "launch",
+            "program": "run.py",
+            "args": [
+                "test_common_unit_auth"
+            ],
+            "console": "integratedTerminal",
+            "justMyCode": true,
+        },
+        {
             "name": "test.filter.unit",
             "type": "debugpy",
             "request": "launch",
