+++ run.py
@@ -558,6 +558,15 @@ class Run:
             ]
         )
 
+    def test_common_unit_auth(self) -> None:
+        Run.set_env_variables(AppType.ALL, AppConfigType.NO_AUTH)
+        pytest.main(
+            Run.DEFAULT_PYTEST_ARGS
+            + [
+                "test/common/unit/auth/",
+            ]
+        )
+
     def test_omopdb_unit(self) -> None:
         Run.set_env_variables(AppType.ALL, AppConfigType.NO_AUTH)
         pytest.main(
